# schema/api/swagger.yaml
swagger: "2.0"
info:
  description: "This is the mplinksters golf club server."
  version: "1.0.0"
  title: "Mariners Point Golf Club"
  termsOfService: "http://mplinksters.club/terms/"
  contact:
    email: "ennead.tbc@gmail.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "www.mplinksters.club"
basePath: "/v1"
...
schemes:
- "https"
- "http"
paths:
  /player:
    post:
      tags:
      - "player"
      summary: "Add a new player to the club"
      description: ""
      operationId: "addPlayer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Player object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Player"
      responses:
        "405":
          description: "Invalid input"
      security:
      - petstore_auth:
        - "write:player"
        - "read:player"
  /player/{playerId}:
    get:
      tags:
      - "player"
      summary: "Find player by ID"
      description: "Returns a single player"
      operationId: "getPlayer"
      produces:
      - "application/json"
      parameters:
      - name: "playerId"
        in: "path"
        description: "ID of player to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/Player"
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Player not found"
      security:
      - api_key: []
    put:
      tags:
      - "player"
      summary: "Updates a player"
      description: "Updates a player"
      operationId: "updatePlayer"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Player object that needs to be added to the store"
        required: true
        schema:
          $ref: "#/definitions/Player"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Player not found"
        "405":
          description: "Invalid input"
      security:
      - petstore_auth:
        - "write:player"
        - "read:player"
    delete:
      tags:
      - "player"
      summary: "Deletes a player"
      description: ""
      operationId: "deletePlayer"
      produces:
      - "application/json"
      parameters:
      - name: "api_key"
        in: "header"
        required: false
        type: "string"
      - name: "petId"
        in: "path"
        description: "Pet id to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "400":
          description: "Invalid ID supplied"
        "404":
          description: "Pet not found"
      security:
      - petstore_auth:
        - "write:pets"
        - "read:pets"
...
securityDefinitions:
  petstore_auth:
    type: "oauth2"
    authorizationUrl: "http://petstore.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write:pets: "modify pets in your account"
      read:pets: "read your pets"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
...